api/requirements.txt
# FastAPI and related dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Database
asyncpg==0.30.0
sqlalchemy>=1.4,<2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pgvector==0.3.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP client
httpx==0.28.1

# Environment
python-dotenv==1.0.1

# External integrations
stripe==7.13.0
resend==2.17.0

# AI/LLM frameworks
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.2.1
langchain-google-genai==2.0.4
langchain-community==0.3.7
openai==1.58.1
anthropic==0.34.2
google-generativeai==0.8.3

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.10.0
isort==5.13.2
flake8==7.1.1
